requirements_airflow.txt
# Core Airflow packages (already included but explicit)
apache-airflow[postgres,redis,celery]==2.7.3
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-redis==3.4.0
apache-airflow-providers-celery==3.5.1
apache-airflow-providers-docker==3.8.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0

# Logging & Monitoring
python-json-logger==2.0.7

# ETL Specific
pyarrow==13.0.0
openpyxl==3.11.0

# CLI utilities
click==8.1.7

# Testing (optional for DAG testing)
pytest==7.4.3
pytest-cov==4.1.0